requirements.txt
numpy==1.26.2
torch==2.1.1
torchaudio==2.1.1
sounddevice==0.4.6
openai-whisper==20231117
openai>=1.3.5
python-dotenv==1.0.0
accelerate>=0.25.0
sentencepiece>=0.1.99
protobuf>=4.25.1
